Designate: Update DB pools configuration (SOC-9767)
Designate pools are interesting as they don't use an API, instead
they use a yaml file to manage pools.

The designate barclamp generates a pool file
`/etc/crowbar/pools.crowbar.yaml`, but because we don't know if a user
has modified this file on purpose, we only generate it if it doesn't
already exist.

Further, we need to be careful when we actaully run:

  designate-manage pool update

Depending on the pools configured in the yaml file, this could nuke any
settings an admin has gone and changed.

As such, this patch is only running the `pool update` once.

There can me multiple pools, but crowbar is only focused on whatever is
defined in pools.crowbar.yaml.

Running the `pool update` once does mean that if things change in the
dns barclamp, these changes would not get reflected in the designate
pools, and so would have to be maintained by the operator.

(cherry picked from commit 850988b76a97f41930044fd868606453db617a4f)