memcache: lookup memcached servers port only on local node (SOC-10173)
With https://github.com/crowbar/crowbar-openstack/pull/2073 merged
we were looking up the port on the node with the lowest mac address,
which however in compilation phase might not yet have that attribute.

By passing in the local node and moving to execution phase we should
always have the default attributes for memcached port visible.