enable LDAP chase_referrals configuration (SOC-7364)
For Microsoft Active Directory deployments where users are not
visible/accessible via the global catalog, list users operation, as a
whole, will fail when one or more referrals are returned on an LDAP
lookup and the current (admin) user have no access to a given referral.

This patch exposes the chase_referrals option in Keystone to enable
customers who got stuck in the above situation to optionally turn
off referrals chasing.

By default, chase_referrals is set to an empty value to preserve
backward compatibility. This is a three state configuration.

Empty value means use the system default. This is all depended on the
underlying LDAP SDK and the system configuration (i.e. /etc/ldap.conf).
"0" means disable referrals chasing.
"1" means enable referrals chasing.

This patch is also exposes the insecure_debug option to help customers and
field engineers to quicky pinpoint the problem when needed.