designate: No longer care about master/slave (SOC-10456)
The mdns recipe was cleaned up to no longer distingush between master
and slave DNS servers. To that end, clean up the code that assigned them
with a simple map, which also handles the case that the dns-server role
was only applied to one node.

Also drive-by correcting the recipe name in the header comment.