Monasca: Handle node reinstall (jsc#SOC-10440, bsc#1148158 )
the chef clinet run after os install during the reinstallation of the
node expects the node is ready with monasca services(eg: moansca-api)
This is false assumption as the monasca-installer (installed on
monasca-master) has not yet run, thus causing disruption as monasca-setup
tries to configure services for monasca-agent.

The monasca-master node has keys which indicate if monasca-installer has
finished successfully. This patch remove those keys when a node with
monasca-server role is being "reinstalled". Thus informing monasca-agent
that monasca services are no longer installed.

This patch does not solve the problem- "re-running moansca-installer"
for the services to be installed. That is a bigger problem as
monasca-master needs the nodes with monasca-server role to be in ready
state with no parallel execution of chef-client.