Fix Cloud 8 no-op migrations (SOC-10623)
Various Cloud 8 proposal migrations have a bug where they
never add the fields they are supposed to add to the
proposal. This is caused by evaluating the proposal hash key
in question as a Boolean truth value the way it happens in
this example:

  unless a["resource_limits"]

This will return an empty string which evaluates to True in a
boolean context. The correct approach is to use Hash.key?().
This commit fixes all occurences of this problem for Cloud 8.