tempest: Fail when enabled_services call failes (SOC-10633)
Sometimes e.g. when changeing the keystone password it seams that
enabled_services is empty this is due to that the openstack cmd call
failes. In any case we should fail if the openstack cmd call fail as
otherwise tempest will set the enabled services to false and we will
actually test nothing.