tempest: don't rely on service catalogue (SOC-10633)
The way that the tempest chef recipe figures out which OpenStack
services can be tested is by inspecting the openstack service
catalogue to retrieve a list of enabled services, instead of
using the barclamp configuration which is more accurate. Worse
even, it does so in the chef compilation phase, before the
running openstack configuration is updated to reflect changes
in the barclamp configuration.

Sometimes, this results in a tempest configuration that doesn't
reflect the list of services that are actuallly deployed. In some of
the more extreme cases, such as updating the admin password, the
OpenStack CLI command used to retrieve the service catalogue
fails completely and returns an empty list, which results in tempest
disabling all services in its configuration.

Inspecting the barclamp role assignments is a more reliable way of
determining which services are actually enabled.

(cherry picked from commit 82eac9d338bdb337a9fc154a1626ed9baa05befb)