Designate: Filter out the admin node (SOC-10658)
The admin node shouldn't be listening on the public network. As such it
can't used as a designate server.

This patch filters it out of the pools list, as crowbar already has a
check to not use the admin node in barclamps that is all we need.

Co-Authored-By: Steve Kowalik

(cherry picked from commit 9a35e8b266e4280c858c8c004380d87c63fd14e5)