No rndc key if no public DNS server (SOC-10835)
In case where DNS server is only running on the admin node, which
is not publicly accessible, we are left with no DNS servers for
Designate integration. For such cases, we do not need to template
the rndc key file used for Bind9 rndc calls.

In this patch we check for such case, if true, we write a warning log
and only template the rndc key file if the case is not valid.

(cherry picked from commit 5f34b2477a95e3b3c3ee2a9956c7912e35f5c53f)