Make sure the input file with ssh key exists (SOC-10133)
Postpone the command into the convergence phase and make
sure it is executed when the input file is there.
Also, postpone the authorized_keys file creation with the use
of lazy evaluation of the variable that was adapted in convergence
phase.

(cherry picked from commit 921e931aea4ce758b3601e465f245a41408fdcf8)