Octavia: account for long ops in HA deployments (SOC-9894)
Installing the amphora image package and creating OpenStack
artifacts (the security group, image, network etc.) can take a lot of
time (between 2 and 2 and a half minutes on ECP Crowbar deployments),
so we have to account for the fact that nodes will fall out of sync in
an HA setup.

We do this by adding an explicit sync with an extended timeout value
at the end of the nova.rb recipe that runs these long operations, to
avoid having timeout failures un subsequent sync marks (the
octavia_database sync mark in our immediate case).