octavia: handle certificate ownership in barclamp (SOC-10906)
The octavia system group and user are created when the
openstack-octavia package is installed on the controller nodes.
This happens automatically when the octavia barclamp is
applied.

However, setting up the paths for the certificates required for
the two-way SSL amphorae communication needs to be done in advance,
before applying the octavia barclamp, and those paths need to have
their ownership set to the octavia system user/group, which aren't
yet available at that time.

This change proposes a solution to this problem, where the admin
can set up those certificate paths on the controller nodes with
root (or other user) as owner, while the Octavia barclamp is
responsible for switching ownership over to the octavia user/group
during application.