neutron: use octavia-api admin VIP URI for lbaasv2 (SOC-10906)
When Octavia is deployed, Neutron is set up to use the LBaaSv2
proxy plugin, which forwards all neutron LBaaSv2 API calls to
the Octavia API service. Neutron was previously misconfigured
to target the API service running locally on the same controller
node instead of targeting the admin VIP, which caused problems
when using a certificate that only includes the VIP IDs.