octavia: switch to noop event streamer (SOC-10868)
With the lbaasv2-proxy neutron service plugin configured,
the Octavia health manager no longer needs the event streamer
to keep the Neutron and Octavia databases in sync, because
Octavia load balancers are only stored in the Octavia database.
In addition to that, the queue_event_streamer has a known issue
with memory leaks (also see [1] and [2]).

[1] https://bugzilla.suse.com/show_bug.cgi?id=1154235
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1631234