designate: declare all mdns servers as master on pool config (SOC-10952)
When creating zones the DNS server notify other servers about the new
zone, however the notification is only accepted when it comes from
masters

Using the VIP as master causes the zone creation to fail as the
notification is refused because it does not come from the VIP, but from
the node IP address. This change fixes it by declaring all mdns server
as master on the default pool defined by crowbar.