tempest: configure Kibana version (SOC-10131)
For Monasca Tempest tests, the Kibana version in use must be
configured for that is the API version the Monasca Tempest
tests use to talk to Kibana (aka the logs-search endpoint). If
this is unconfigured, the tempest tests will use an out of
date version which causes the requests they issue to get
rejected by Kibana. This commit configures the Kibana version
to whatever is the most recent Kibana version available in
zypper, thus ensuring it matches what is installed or gets
installed on the monasca-server node.