octavia: configure barbican auth (SOC-10989)
Even though Octavia uses the service catalogue to auto-configure
the Barbican service endpoint, it helps if we use an endpoint
hard-coded in the configuration files because there is no other
way to restart Octavia services in case the Barbican endpoint
also changes (e.g. switch from HTTP to HTTPs or vice-versa).

This commit also sets the insecure option not just for Barbican,
but also for other services that Octavia interacts with (Nova
and Neutron).