octavia: enable octavia tempest plugin test cases (SOC-8743)
Add a tempest test run filter for the octavia tempest plugin
test cases and adjust the tempest configuration for it.

The `load-balancer_member` still needs to be added by tempest
for the legacy neutron-lbaas tempest test cases to pass, while
the octavia tempest test cases handle load balancer roles
internally. This creates a conflict situation which is solved
by configuring the `RBAC_test_type` octavia tempest option
differently, depending on whether the legacy neutron lbaasv2
API is enabled and used as a proxy for the Octavia API.