rabbitmq: sync startup definitions.json with recipe (SOC-11077,SOC-11274)
In an HA setup, every time a rabbitmq broker is restarted, whatever the
reason may be (an RA monitor failure, a controller node reboot etc.), the
rabbitmq queue policy is changed from this:

openstack       ha-queues       queues  ^(?!(amq.)|(.*_fanout_)|(reply_)).*     {"ha-mode":"exactly","ha-params":2,"ha-sync-mode":"automatic"}  0

to this:

openstack       ha-queues       queues  ^(?!amq.).*     {"ha-mode":"exactly","ha-params":2}     0

This behavior isn't intentional, it's just a byproduct of neglect (the
startup rabbitmq configuration has over time fallen behind the runtime
configuration enforced by the chef recipes). It has a negative impact
on performance and reliability, because it abruptly makes fanout and
reply queues fully mirrored, making matters worse in the middle of a
restart operation that already puts a lot of strain on the rabbitmq
cluster.

The policy in the generated /etc/rabbitmq/definitions.json
file used during startup needs to be kept in sync with the policy
configured at runtime by the rabbitmq chef recipe, to avoid
inconsistencies. This is what this change is all about.

The policy configuration is kept in the CrowbarRabbitmqHelper
and invoked both to generate the definitions.json file and to
configure the rabbitmq policy.

This change should improve the rabbitmq cluster reliability during
events such as broker restarts or controller node reboots.