Designate: make sure dns-server is active on a non-admin node (SOC-10636)
Designate uses Bind9 backend by default. Bind9 is managed by the
dns-server role. Since Designate does not rely on the DNS server
on the Crowbar admin node, we need to make sure DNS server is running
on at least one non-admin node. If HA is enabled, DNS server should be
running on all the Designate worker nodes. If not, we need to warn users
the need to do so.