keystone: fix keystone node lookup (SOC-11333, bsc#1164838)
In a cloud where a controller was added after initial
deployment and MAC addresses are out of sequence, the
search_for_keystone() method may return the wrong controller
in the first upgraded controller's crowbar_join phase. This
commit ensures this does not happen by restricting the search for
nodes that are not in state `crowbar_upgrade`.