octavia: move management network creation to octavia barclamp (SOC-10904)
Move the logic creating the Octavia management network away from
the neutron barclamp and into the octavia barclamp, to avoid
situations where the network isn't created because the neutron
create_default_networks attribute is explicitly disabled, as
well as to centralize all the logic responsible for creating
octavia prerequisites to the octavia barclamp. This breaks the
circular dependency between Neutron and Octavia (e.g. the name
of the management network is coded as an Octavia barclamp
attribute, which means Neutron would require Octavia to be
deployed to have access to this attribute, while Octavia relies
on this network to be already created by Neutron before its
services can be configured).