neutron: don't add physnets for non-enabled networks (SOC-11204)
Deploying octavia on top of an already deployed cloud and using
the `octavia` network.json network consists of the following steps:
 1. add the `octavia` network to the network barclamp (raw mode) and
 apply the network barclamp. This will set up proper interfaces and
 bridges for the `octavia` network
 2. re-apply neutron, then nova barclamps, to update the neutron
 server and agent configurations with the proper physnet mappings
 3. finally, apply the octavia barclamp

It was discovered that during step 1., the neutron recipes incorrectly
associate the `octavia` network with the physnet1 mapping, because
the function that computes the physnet mappings calls `get_network_by_type`
and that function (mysteriously) returns the admin network if the network
being looked up isn't enabled on the node.

This commit fixes this problem (only) for octavia, by double-checking that
the octavia network is actually enabled on the node before taking
it into consideration. However, the problem might also apply to
other networks (ironic, additional_external_networks ?) when updating the
list of flat networks in the network-server ML2 configuration on cloud
deployments that separate the neutron network nodes from the control nodes.