octavia: remove mgmt_net from UI (SOC-10904)
The recommended way of configuring the Octavia management network
is by providing an `octavia` network in the network barclamp configuration,
while the Octavia barclamp takes care of the rest. This means it makes
no sense to keep the option of naming a different neutron provider network
as a main configuration option in the UI, because it may confuse customers
into thinking they still need to provide this network themselves.

The option will still be accessible through the raw barclamp view, if
necessary.