tempest: retry openstack commands (SOC-11238)
On a HA deployment, actions that requires reapplying barclamps for
openstack services (e.g. enabling SSL on keystone) causes the tempest
barclamp to occasionally fail.

It fails because while reapplying the tempest barclamp
(deployed in one of the controllers), on other nodes everything
was done and the post actions are called which triggers the restart
of openstack services, when the openstack commands from the tempest
barclamp are called at the same time as the service is restarting
those calls fails.

This change adds a retry on the ruby blocks executing openstack
commands. So if the command fails it tries again, giving time for the
service to restart.

(cherry picked from commit 50d3409abba4b5e319e7d3f7626fda47f1ee848d)