nova: run keystone_register on cluster founder only (SOC-11243)
The execution of the nova cookbook includes restarting apache,
however, when apache is restarted while keystone_register is running on
the non-cluster founder nodes it fails as keystone (running under apache)
is also being restarted.

This change fixes that by including a condition for running the
keystone_register operations only on the cluster founder node.