run keystone_register on cluster founder only when HA (SOC-11248)
Although it was fixed for some services (e.g. nova ) there are still
other services where keystone_register is called on all controllers with
a subsequent apache restart. From this, there is a risk of the
keystone_register call be executed while apache (and keystone) is
restarting and consequently failing the deployment.

This change make all calls for keystone_register to be made only on
one of the controllers for all services that supports HA.