[4.0] Fix create magnum k8s image and flavor (SOC-11251)
The conditional to check if the image already exists its not working
properly. There are cases where the openstack command is executed while
the openstack services are restarting causing the command to fail. When
the command fails it means that the image does not exist (which is not
the case) and the image is created again.

This changes decouple the conditional checking of if the image/flavor
exists by moving it to a dedicated ruby block. If for some reason the
command fails the barclamp will also fail.