Disable keepalive in Keystone Apache config (bsc#1179084)
Horizon's cache is preventing the garbage collector from removing
objects [1] with connections to Keystone. Keystone is running under
Apache which closes these connections after 15 seconds and they move
to CLOSE_WAIT on the client side. This can lead to the Horizon process
running out of file descriptors.

This patch disables KeepAlive in the Keystone Apache config, so the
server sends "Connection: close" in the HTTP response header and the
client removes the connection immediately.

[1] https://bugs.launchpad.net/horizon/+bug/1793411