neutron: Fix handling of networks with non-ascii names (SOC-11429)
Chef uses default LC_ALL=C locale for all internal ShellOut calls.
These fails if command output contains non-ASCII characters.
The problem is bigger but it was really fixed only in Chef 12.
We could patch ShellOut to use different default be it could cause
some unwanted side effects in many places.
In this case the problem is with network names so it's easier to
limit the output to include only expected entries and not print all
names.