keystone wakeup: get new session on any error (bsc#1189052)
Before f0f3e481 the keystone wakeup action would always get a
new token when it started, but now it tries to reuse the session
object. The sessions's token will eventually expire which will
cause an authentication error, but the retry loop only tries to
get a new session/token if the password changes. This patch
changes the retry loop so it gets a new session on any error.